requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
black==25.1.0
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.1
django-debug-toolbar==6.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.29.0
flake8==7.3.0
frozenlist==1.8.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
mccabe==0.7.0
multidict==6.7.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
stripe==14.3.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
yarl==1.22.0
psycopg2-binary==2.9.11
psycopg==3.3.2
redis==6.4.0
